added default settings and cleanup the write_job_files script

+++ job_template.sh
@@ -7,20 +7,22 @@
 #PBS -q long
 #PBS -S /cvmfs/icecube.opensciencegrid.org/py2-v2/icetray-start
 FINAL_OUT={final_out}
+KEEP_CRASHED_FILES={keep_crashed_files:d}
 echo $FINAL_OUT
 if [ -z ${PBS_JOBID} ] && [ -z ${CLUSTER} ]
 then
     echo 'Running Script w/o temporary scratch'
     {script_folder}/step_{step_number}.py {yaml_copy} {run_number} --no-scratch
     echo 'IceTray finished with Exit Code: ' $?
-    if [ "$?" != "0" ] ; then
+    if [ "$?" != "0" ] && [ !$KEEP_CRASHED_FILES ] ; then
+        echo 'Deleting partially processed file!'
         rm $FINAL_OUT
     fi
 else
     echo 'Running Script w/ temporary scratch'
     {script_folder}/step_{step_number}.py {yaml_copy} {run_number} --scratch
     echo 'IceTray finished with Exit Code: ' $?
-    if [ "$?" = "0" ] ; then
+    if [ "$?" = "0" ] || && [ !$KEEP_CRASHED_FILES ]; then
         cp {scratch_out} $FINAL_OUT
     else
         rm {scratch_out}
